Save the power law reference frequency to the outputfile and use it for plotting.
